fix: empty graph is now fixed
